+++ src/main/java/br/edu/fateczl/Hotel/model/dto/ReservaDTO.java
@@ -1,7 +1,5 @@
 package br.edu.fateczl.Hotel.model.dto;
 
-import java.math.BigInteger;
-
 import br.edu.fateczl.Hotel.model.dto.interfaces.IDTO;
 import br.edu.fateczl.Hotel.model.entity.Pessoa;
 import br.edu.fateczl.Hotel.model.entity.Quarto;
@@ -9,7 +7,7 @@ import br.edu.fateczl.Hotel.model.entity.Reserva;
 import br.edu.fateczl.Hotel.model.entity.Vaga;
 
 public class ReservaDTO implements IDTO<Reserva> {
-	private BigInteger id;
+	private Long id;
 	private String dataInicio;
 	private String dataFim;
 	private PessoaDTO documento;
@@ -17,11 +15,11 @@ public class ReservaDTO implements IDTO<Reserva> {
 	private VagaDTO vaga;
 
 	
-	public BigInteger getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(BigInteger id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
