+++ src/main/java/br/edu/fateczl/Hotel/repository/ReservaRepository.java
@@ -1,6 +1,5 @@
 package br.edu.fateczl.Hotel.repository;
 
-import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -10,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import br.edu.fateczl.Hotel.model.entity.Reserva;
 
 @Repository
-public interface ReservaRepository extends JpaRepository<Reserva, BigInteger>{
+public interface ReservaRepository extends JpaRepository<Reserva, Long>{
 	@Query(
 			value = "SELECT "
 					+ "	* "
